@@ -41,7 +41,7 @@
     "@types/marked": "^2.0.5",
     "@types/mockjs": "^1.0.9",
     "@types/node": "^20.8.7",
-    "@types/react": "^18.2.33",
+    "@types/react": "^18.0.17",
     "@types/react-datepicker": "^4.19.1",
     "@types/react-dom": "^18.2.14",
     "@types/styled-components": "^5.1.29",
@@ -63,7 +63,7 @@
     "style-loader": "^3.3.3",
     "terser-webpack-plugin": "^5.3.9",
     "ts-loader": "^9.5.0",
-    "typescript": "^4.9.5",
+    "typescript": "^5.2.2",
     "webpack": "^5.89.0",
     "webpack-cli": "^4.10.0",
     "webpack-dev-server": "^3.11.3"
