@@ -1,7 +1,7 @@
 import React from 'react';
 import styled from 'styled-components';
 import { useNavigate } from 'react-router-dom';
-import { Button } from '@/components';
+import { Button } from '@/components/button';
 import { IPost } from '@/types';
 import { fetchPosts, deletePost } from '@/apis/posts';
 
